knowledge-graph-builder/fake_news/requirements.txt
backcall==0.1.0
beautifulsoup4==4.8.0
blis==0.2.4
boto3==1.9.224
botocore==1.12.224
certifi==2019.6.16
chardet==3.0.4
cssselect==1.1.0
cymem==2.0.2
decorator==4.4.0
docutils==0.15.2
feedfinder2==0.0.4
feedparser==5.2.1
idna==2.8
ipdb==0.12.2
ipython==7.7.0
ipython-genutils==0.2.0
isodate==0.6.0
jedi==0.15.1
jieba3k==0.35.1
jmespath==0.9.4
jsonschema==2.6.0
lxml==4.4.1
murmurhash==1.0.2
neo4j==1.7.4
neobolt==1.7.13
neotime==1.7.4
neuralcoref==4.0
newsapi-python==0.2.5
newspaper3k==0.2.8
nltk==3.4.5
numpy==1.17.1
parso==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.1.0
plac==0.9.6
preshed==2.0.1
prompt-toolkit==2.0.9
ptyprocess==0.6.0
Pygments==2.4.2
pyparsing==2.4.2
python-dateutil==2.8.0
pytz==2019.2
PyYAML==5.1.2
rdflib==4.2.2
requests==2.21.0
requests-file==1.4.3
s3transfer==0.2.1
six==1.12.0
soupsieve==1.9.3
spacy==2.1.3
SPARQLWrapper==1.8.4
srsly==0.1.0
thinc==7.0.8
tinysegmenter==0.3
tldextract==2.2.1
tqdm==4.35.0
traitlets==4.3.2
urllib3==1.24.3
wasabi==0.2.2
wcwidth==0.1.7
